+++ pkg/cmd/root.go
@@ -1,7 +1,10 @@
 package cmd
 
 import (
+	"context"
+	"errors"
 	"os"
+	"time"
 
 	"github.com/cyberark/conjur-cli-go/pkg/version"
 	"github.com/spf13/cobra"
@@ -16,6 +19,7 @@ func newRootCommand() *cobra.Command {
 	}
 
 	rootCmd.PersistentFlags().BoolP("debug", "d", false, "Debug logging enabled")
+	rootCmd.PersistentFlags().Duration("timeout", time.Minute, "HTTP timeout duration, between 1s and 10m")
 	rootCmd.SetVersionTemplate("Conjur CLI version {{.Version}}\n")
 	return rootCmd
 }
@@ -26,6 +30,10 @@ func Execute() {
 	rootCmd.SetOut(os.Stdout)
 	rootCmd.SetErr(os.Stderr)
 	err := rootCmd.Execute()
+	if errors.Is(err, context.DeadlineExceeded) {
+		rootCmd.PrintErrln(
+			"Your request has timed out. If your operation is expected to be long-running, please consider increasing the HTTP timeout. For details, please refer to the command help.")
+	}
 	if err != nil {
 		os.Exit(1)
 	}